Ports https://github.com/tgstation/tgstation/pull/57825: Improve tgui routing to allow interface subdirectories; apply to CyborgModuleRewriter
