+++ tgui/packages/tgui/interfaces/CyborgModuleRewriter/ModuleView/index.js
@@ -5,15 +5,15 @@
  * @license ISC
  */
 
-import { Flex, Section, Tabs, Button } from '../../../components';
-import EmptyPlaceholder from '../EmptyPlaceholder';
-import Module from './Module';
+import { Button, Flex, Section, Tabs } from '../../../components';
+import { EmptyPlaceholder } from '../EmptyPlaceholder';
+import { Module } from './Module';
 
 // width hard-coded to allow display of widest current module name
 // without resizing when ejected/reset
 const ModuleListWidth = 18;
 
-const ModuleView = props => {
+export const ModuleView = props => {
   const {
     modules: {
       available = [],
@@ -112,5 +112,3 @@ const ModuleView = props => {
       )
   );
 };
-
-export default ModuleView;
