+++ tgui/packages/tgui/routes.js
@@ -8,7 +8,7 @@ import { selectBackend } from './backend';
 import { selectDebug } from './debug/selectors';
 import { Window } from './layouts';
 
-const requireInterface = require.context('./interfaces', false);
+const requireInterface = require.context('./interfaces');
 
 const routingError = (type, name) => () => {
   return (
@@ -47,22 +47,27 @@ export const getRoutedComponent = store => {
     }
   }
   const name = config?.interface;
+  const interfacePathBuilders = [
+    name => `./${name}.tsx`,
+    name => `./${name}.js`,
+    name => `./${name}/index.tsx`,
+    name => `./${name}/index.js`,
+  ];
   let esModule;
-  try {
-    esModule = requireInterface(`./${name}.tsx`);
-  }
-  catch (err) {
-    if (err.code === 'MODULE_NOT_FOUND') {
-      try {
-        esModule = requireInterface(`./${name}.js`);
-      }
-      catch (err) {
-        if (err.code === 'MODULE_NOT_FOUND') {
-          return routingError('notFound', name);
-        }
+  while (!esModule && interfacePathBuilders.length > 0) {
+    const interfacePathBuilder = interfacePathBuilders.shift();
+    const interfacePath = interfacePathBuilder(name);
+    try {
+      esModule = requireInterface(interfacePath);
+    }
+    catch (err) {
+      if (err.code !== 'MODULE_NOT_FOUND') {
         throw err;
       }
     }
+  }
+  if (!esModule) {
+    return routingError('notFound', name);
   }
   const Component = esModule[name];
   if (!Component) {
